+++ src/mcmc/gibbs.jl
@@ -306,6 +306,26 @@ end
 
 get_varinfo(state::GibbsState) = state.vi
 
+function check_all_variables_handled(vns, spl::Gibbs)
+    handled_vars = Iterators.flatten(spl.varnames)
+    missing_vars = [
+        vn for vn in vns if !any(hv -> AbstractPPL.subsumes(hv, vn), handled_vars)
+    ]
+    if !isempty(missing_vars)
+        msg =
+            "The Gibbs sampler does not have a component sampler for: $(join(missing_vars, ", ")). " *
+            "Please assign a component sampler to each variable in the model."
+        throw(ArgumentError(msg))
+    end
+end
+
+function Turing._check_model(model::DynamicPPL.Model, spl::Gibbs)
+    # TODO(penelopeysm): Could be smarter: subsamplers may not allow discrete variables.
+    Turing._check_model(model, !Turing.allow_discrete_variables(spl))
+    varnames = keys(rand(model))
+    return check_all_variables_handled(varnames, spl)
+end
+
 function AbstractMCMC.step(
     rng::Random.AbstractRNG,
     model::DynamicPPL.Model,